Move IssueCard to main ui package
